requirements.txt
telethon~=1.38.1
openai~=1.59.6
python-dotenv~=1.0.1
pymongo~=4.10.1